requirements.txt
# Core server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.6.0

# MCP protocol
mcp>=1.0.0

# Memory engine
chromadb>=0.4.24
networkx>=3.2.1
anthropic>=0.25.0

# Embeddings (local)
ollama>=0.1.8

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
